+++ packages/policies/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added DomainPolicy to filter events by NIP-05 domain.
+
 ## 0.33.1 - 2024-09-11
 
 ### Fixed